requirements.txt
cffi==1.16.0
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.2
ids-peak 
ids-peak-ipl 
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pyAndorSDK2 
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
scipy==1.11.4
six==1.16.0
tifffile==2023.12.9
wxPython==4.2.1
